Update symlinks after compressing the files they point to
We compress the underlying firmware files, renaming them with the .zst
extension. We need to update the symlinks to point to the new targets,
and rename the symlinks themselves.

Some of the symlinks in linux-firmware-wifi depend on firmware files in
linux-firmware, so add a package dependency.

See https://github.com/clearlinux/distribution/issues/2779